Logically split code and add more documentation
