+++ tests/test_unit_queue.py
@@ -1,4 +1,5 @@
 import time, os, random
+from csv import reader
 from pytest import *
 from dataclasses import dataclass
 from expertise.service.queue import *
